requirements.txt
fastapi
uvicorn
requests
python-dotenv
docxtpl
python-docx
